requirements.txt
aiohttp==1.0.5
async-timeout==1.3.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
discord.py==0.16.11
html5lib==0.999999999
idna==2.6
imgurpython==1.1.7
multidict==3.1.3
psycopg2==2.7.3.1
pycparser==2.18
recordtype==1.1
requests==2.18.4
six==1.10.0
urllib3==1.22
webencodings==0.5.1
websockets==3.4
youtube-dl==2017.9.24
